+++ javascript/UnixLepik.d.ts
@@ -89,6 +89,7 @@ declare class UnixLepik {
      * @returns {number} The ID of the active window.
      */
     getActiveWindowId(): number;
+    setActiveWindow(windowId: string): void;
     /**
      * Delays the execution for the specified number of milliseconds.
      * @param {number} ms - The number of milliseconds to delay.