+++ tests/mouseMove.js
@@ -1,7 +1,8 @@
 const lepik = require("../index.js")
 
 async function main() {
-  lepik.paste()
+  lepik.delay(2000)
+  lepik.setActiveWindow(2426196)
   lepik.close()
 }
 function sleep(ms) {