+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 [Full Changelog](https://github.com/borecjeborec1/LepikJS/commits/main)
 
 ##### Implemented enhancements:
+-  setActiveWindow method ([#7cd3529](https://github.com/borecjeborec1/LepikJS/commit/7cd3529fb8f79dbcc6c8435bbfa38b1e06e3301d))
 -  delay functionality ([#c576ed3](https://github.com/borecjeborec1/LepikJS/commit/c576ed3e53bb839e3e2a85a7eb270178ce36f313)) 
 
 ##### Fixed bugs: